+++ jest.config.js
@@ -15,24 +15,24 @@ module.exports = {
   coverageDirectory: "coverage",
 
   // An object that configures minimum threshold enforcement for coverage results
-  /* coverageThreshold: {
+  coverageThreshold: {
     global: {
       branches: 100,
       functions: 100,
       lines: 100,
-      statements: 100
-    }
-  }, */
+      statements: 100,
+    },
+  },
 
   // The test environment that will be used for testing
-  testEnvironment: "node",
+  testEnvironment: "jsdom",
 
   // The glob patterns Jest uses to detect test files
   testMatch: ["**/test/**/?(*.)+(spec|test).js?(x)"],
 
   transform: {
-    ".js$": "babel-jest"
+    ".js$": "babel-jest",
   },
 
-  setupFiles: ["<rootDir>/jest.init.js"]
+  setupFiles: ["<rootDir>/jest.init.js"],
 };